\name{createAppCmd}
\alias{createAppCmd}
\title{createAppCmd}
\usage{
  createAppCmd(cmdName, params = NULL, delim = "&")
}
\arguments{
  \item{cmdName}{}

  \item{params}{}

  \item{delim}{}
}
\value{
  the command with delemited params.
}
\description{
  createAppCmd
}
\author{
  Barnet Wagman
}
\keyword{internal}

