\name{createAppConnectCommand}
\alias{createAppConnectCommand}
\title{sendAppCmd}
\usage{
  createAppConnectCommand(hostName, userName, portRJ,
    portJJ, pemFile)
}
\arguments{
  \item{hostName}{}

  \item{userName}{}

  \item{portRJ}{}

  \item{portJJ}{}

  \item{pemFile}{return con}
}
\description{
  sendAppCmd
}
\author{
  Barnet Wagman
}
\keyword{internal}

