\name{createExprMessage}
\alias{createExprMessage}
\title{Creates an rcs message that contains an expression for remote
evaluation.  Note that the message createe by this funtion
includes maxBytesReturnable element.}
\usage{
  createExprMessage(expr, maxBytesReturnable = NULL)
}
\arguments{
  \item{type}{}

  \item{maxBytesReturnable}{}
}
\value{
  message
}
\description{
  Creates an rcs message that contains an expression for
  remote evaluation.  Note that the message createe by this
  funtion includes maxBytesReturnable element.
}
\author{
  Barnet Wagman
}
\keyword{internal}

