\name{createSerializedMessage}
\alias{createSerializedMessage}
\title{Create serialized message}
\usage{
  createSerializedMessage(type, val, so)
}
\arguments{
  \item{type}{}

  \item{val}{}

  \item{maxBytesReturnable}{}
}
\value{
  serialized message
}
\description{
  Create serialized message
}
\author{
  Barnet Wagman
}
\keyword{internal}

