\name{evalExpression}
\alias{evalExpression}
\title{evalExpression}
\usage{
  evalExpression(con, expr, maxBytesReturnable)
}
\arguments{
  \item{con}{}

  \item{expr}{}

  \item{maxBytesReturnable}{}
}
\value{
  value
}
\description{
  evalExpression
}
\author{
  Barnet Wagman
}
\keyword{internal}

