\name{getJavaAppClassPath}
\alias{getJavaAppClassPath}
\title{getJavaAppClassPath
This function creates an expression that java will expand to
a list of all jar in the package 'jars' directory.
quotes.}
\usage{
  getJavaAppClassPath(packageName, quoted = TRUE)
}
\arguments{
  \item{packageName}{}
}
\value{
  Java classpath as character.
}
\description{
  getJavaAppClassPath This function creates an expression
  that java will expand to a list of all jar in the package
  'jars' directory. quotes.
}
\author{
  Barnet Wagman
}
\keyword{internal}

