\name{getJavaAppClassPath.v0}
\alias{getJavaAppClassPath.v0}
\title{getJavaAppClassPath.v0
This function gets the names of all jars in the 'jars' directory
of the specified package}
\usage{
  getJavaAppClassPath.v0(packageName)
}
\arguments{
  \item{packageName}{}
}
\value{
  Java classpath as character.
}
\description{
  getJavaAppClassPath.v0 This function gets the names of
  all jars in the 'jars' directory of the specified package
}
\author{
  Barnet Wagman
}
\keyword{internal}

