\name{getMaxBytesReturnable}
\alias{getMaxBytesReturnable}
\title{getMaxBytesReturnable}
\usage{
  getMaxBytesReturnable(m,
    maxBytesReturnableDefault = 2^16)
}
\arguments{
  \item{m}{rcs message of type EXPR}
}
\value{
  maximum size (in bytes) of a returnable object.
}
\description{
  getMaxBytesReturnable
}
\author{
  Barnet Wagman
}
\keyword{internal}

