\name{hasEl}
\alias{hasEl}
\title{Test for list with specfied named element,}
\usage{
  hasEl(x, name)

  hasEl(x, name)
}
\arguments{
  \item{x}{}

  \item{name}{}

  \item{x}{}

  \item{name}{}
}
\value{
  boolean

  boolean
}
\description{
  Test for list with specfied named element,

  Test for list with specfied named element,
}
\author{
  Barnet Wagman

  Barnet Wagman
}
\keyword{internal}

