\name{isOpenEh}
\alias{isOpenEh}
\title{isOpen() sometimes throws 'invalid connection' error: this wrapper
returns FALSE when an error occurs.}
\usage{
  isOpenEh(con)
}
\arguments{
  \item{con}{}
}
\value{
  boolean
}
\description{
  isOpen() sometimes throws 'invalid connection' error:
  this wrapper returns FALSE when an error occurs.
}
\author{
  Barnet Wagman
}
\keyword{internal}

