\name{readMessage}
\alias{readMessage}
\title{Reads a message, waiting the specified time, possibly forever.}
\usage{
  readMessage(con, maxWaitSecs = -1)
}
\arguments{
  \item{con}{}

  \item{maxWaitSecs}{}
}
\value{
  the message obj.
}
\description{
  Reads a message, waiting the specified time, possibly
  forever.
}
\author{
  Barnet Wagman
}
\keyword{internal}

