\name{rre.closeAllConnections}
\alias{rre.closeAllConnections}
\title{Close all connections}
\usage{
  rre.closeAllConnections()
}
\description{
  Closes all connections to rreval servers and terminates
  the Java application that handles these connections
  (rreval.RReClientApp).
}
\author{
  Barnet Wagman
}

