\name{rre.showClients}
\alias{rre.showClients}
\title{Show clients}
\usage{
  rre.showClients()
}
\value{
  (hostName,userName) pairs in a data.frame.
}
\description{
  Displays a list of active clients in this R session, i.e.
  the names of rre server hosts to which there are open
  connections.
}
\author{
  Barnet Wagman
}

