\name{sendEvalResult}
\alias{sendEvalResult}
\title{sendEvalResults}
\usage{
  sendEvalResult(con, val, maxBytesReturnable, so = NULL)
}
\arguments{
  \item{con}{}

  \item{val}{}

  \item{maxBytesReturnable}{}

  \item{so}{}
}
\description{
  sendEvalResults
}
\author{
  Barnet Wagman
}
\keyword{internal}

