\name{sendMessageToRReServer}
\alias{sendMessageToRReServer}
\title{accessAppIsRunning}
\usage{
  sendMessageToRReServer(con, messageRcs, maxWaitSecs = -1)
}
\arguments{
  \item{con}{}

  \item{messageRcs}{}

  \item{maxWaitSecs}{}
}
\description{
  accessAppIsRunning
}
\author{
  Barnet Wagman
}
\keyword{internal}

