\name{writeMessage}
\alias{writeMessage}
\title{writeMessage}
\usage{
  writeMessage(con, messageType, obj, nMaxTries = 16,
    maxWaitForAckSecs = 2)
}
\arguments{
  \item{con}{}

  \item{messageType}{}

  \item{obj}{return number of chars written.}
}
\description{
  writeMessage
}
\author{
  Barnet Wagman
}
\keyword{internal}

