\name{writeString}
\alias{writeString}
\title{writeString}
\usage{
  writeString(con, s, messageType, maxWaitForAckSecs = 3)
}
\arguments{
  \item{con}{}

  \item{s}{}

  \item{messageType}{}

  \item{maxWaitForAcks}{}
}
\value{
  the number of chars written, if successful, else an error
  is thrown.
}
\description{
  writeString
}
\author{
  Barnet Wagman
}
\keyword{internal}

