\name{getJavaAppClassPath}
\alias{getJavaAppClassPath}
\title{getRRevalJavaAppClassPath
This function gets the names of all jars in the 'jars' directory
of the specified package}
\usage{
  getJavaAppClassPath(packageName)
}
\arguments{
  \item{packageName}{}
}
\value{
  Java classpath as character.
}
\description{
  getRRevalJavaAppClassPath This function gets the names of
  all jars in the 'jars' directory of the specified package
}
\author{
  Barnet Wagman
}
\keyword{internal}

